主题配置合并入系统配置单元
