+++ src/main/resources/config/setting/setting.properties
@@ -1,2 +1,6 @@
 #设置是否强制启用实名认证
 setting.authentication.force=0
+#设置是否使用黑白网页
+theme.grayscale=0
+#设置是否使用Pride主题
+theme.pride=0
