--- src/main/resources/config/theme/theme.properties
@@ -1,4 +1,0 @@
-#设置是否使用黑白网页
-theme.grayscale=0
-#设置是否使用Pride主题
-theme.pride=0
